Refactor OpenAI config and handler initialization
Moved OpenAI usage limits to a dedicated LIMITS class in Settings and added BASE_URL support for OpenAI API. Refactored GptHandler to lazily initialize OpenAI and Google clients with support for custom base URLs. Updated OpenAiEmbedding to use model name in database key. Adjusted AdaEmbedding to use float embeddings. Updated ProviderManager to use new LIMITS config and commented out Copilot and process-based provider registration. Added GuildGPTHandler stub.

+++ src/main/java/link/locutus/discord/gpt/imps/AdaEmbedding.java
@@ -38,10 +38,10 @@ public class AdaEmbedding extends AEmbeddingDatabase {
         if (data.size() != 1) {
             throw new RuntimeException("Expected 1 embedding, got " + data.size());
         }
-        List<Double> result = data.get(0).embedding();
+        List<Float> result = data.get(0).embedding();
         float[] target = new float[result.size()];
         for (int i = 0; i < target.length; i++) {
-            target[i] = result.get(i).floatValue();
+            target[i] = result.get(i);
         }
         return target;
     }
